# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lettura API",
    "description": "Lettura is a wallabag-inspired read-it-later service. Endpoints below are stable; unannotated routes exist but their schema is not yet machine-readable.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tokens": {
      "get": {
        "tags": [
          "tokens"
        ],
        "operationId": "list_tokens",
        "responses": {
          "200": {
            "description": "List of personal access tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalAccessToken"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Token management requires interactive login"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tokens"
        ],
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Token created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Token management requires interactive login"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tokens/{id}": {
      "delete": {
        "tags": [
          "tokens"
        ],
        "operationId": "delete_token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Token ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token deleted"
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Token management requires interactive login"
          },
          "404": {
            "description": "Token not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTokenRequest": {
        "type": "object",
        "required": [
          "name",
          "scope"
        ],
        "properties": {
          "expires_in_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        }
      },
      "CreateTokenResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "scope",
          "token"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "description": "Plaintext token — only returned on creation, never again."
          }
        }
      },
...
      "PersonalAccessToken": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "token_prefix",
          "scope",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "token_prefix": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}